requirements.txt
apispec==1.3.3
argcomplete==1.11.1
attrs==19.3.0
Babel==2.8.0
backcall==0.1.0
bleach==3.1.3
cached-property==1.5.1
cattrs==0.9.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
colorlog==4.0.2
configparser==3.5.3
croniter==0.3.31
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.1.1
docutils==0.16
entrypoints==0.3
flake8==3.7.9
funcsigs==1.0.2
graphviz==0.13.2
gunicorn==19.10.0
importlib-metadata==1.5.0
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
iso8601==0.1.12
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.10.3
joblib==0.15.1
json-merge-patch==0.2
json5==0.9.3
jsonschema==3.2.0
jupyter-client==6.1.0
jupyter-core==4.6.3
jupyterlab==2.0.1
jupyterlab-server==1.0.7
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
llvmlite==0.31.0
lockfile==0.12.2
luigi==2.8.12
Markdown==2.6.11
MarkupSafe==1.1.1
marshmallow==2.19.5
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.22.2
matplotlib==3.2.1
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
numba==0.48.0
numexpr==2.7.1
numpy==1.18.1
opencv-python==4.2.0.32
pandas==0.25.3
pandocfilters==1.4.2
parso==0.6.2
pendulum==1.4.4
pickleshare==0.7.5
Pillow==7.0.0
plotly==4.5.4
prison==0.1.2
prometheus-client==0.7.1
prompt-toolkit==3.0.4
psutil==5.7.0
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.6.1
PyJWT==1.7.1
pyparsing==2.4.6
pyradiomics==3.0
pyrsistent==0.15.7
python-daemon==2.1.2
python-dateutil==2.8.1
python-editor==1.0.4
python3-openid==3.1.0
pytz==2019.3
pytzdata==2019.3
PyYAML==5.3
pyzmq==19.0.0
requests==2.23.0
retrying==1.3.3
scikit-image==0.16.2
scikit-learn==0.22.2.post1
scipy==1.4.1
seaborn==0.10.0
Send2Trash==1.5.0
setproctitle==1.1.10
SimpleITK==1.2.4
six==1.14.0
tables==3.6.1
tabulate==0.8.6
tenacity==4.12.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.2
thrift==0.13.0
tornado==5.1.1
tqdm==4.43.0
traitlets==4.3.3
tzlocal==1.5.1
unicodecsv==0.14.1
urllib3==1.25.8
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==0.16.1
zipp==3.1.0
